+++ src/app/store/dialogs/login-form/store-login-form-dialog.component.ts
@@ -10,7 +10,7 @@ import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms'
 import { MatDialogRef } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { BehaviorSubject, Observable, Subject, throwError } from 'rxjs';
-import { catchError, map, switchMap, tap } from 'rxjs/operators';
+import { catchError, map, switchMap, takeUntil, tap } from 'rxjs/operators';
 import { AuthenticationService } from 'src/app/authentication.service';
 import { ProfileService } from 'src/app/profile.service';
 import { DialogSwitcherButtonComponent } from 'src/app/shared/components/dialog-switcher-button/dialog-switcher-button.component';
@@ -75,6 +75,7 @@ export class StoreLoginFormDialogComponent
       };
 
       this.authenticationService.login(details).pipe(
+        takeUntil(this.authenticationService.authCancelation$),
         tap(() => {
           this.dialog.close();
           const successMessage = $localize`:Message of success after login:You have logged in`